Frontend integration & Refactoring
