+++ frontend/src/components/habits/HabitList.css
@@ -1,0 +1,173 @@
+/* ============================================================
+   HABIT LIST
+   ============================================================ */
+
+.task-row {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  padding: 10px 10px;
+  border-radius: var(--radius-sm);
+  transition: background var(--dur-fast);
+  position: relative;
+}
+
+.task-row:hover {
+  background: var(--surface-2);
+}
+
+/* Mark-as-done button */
+.task-done-btn {
+  width: 28px;
+  height: 28px;
+  border-radius: 8px;
+  border: 1.5px solid var(--border);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  transition: all var(--dur-fast) var(--ease);
+  color: var(--text-3);
+  background: transparent;
+}
+
+.task-done-btn:hover {
+  border-color: var(--accent-border);
+  color: var(--accent);
+  background: var(--accent-glow);
+}
+
+.task-done-btn.done {
+  background: var(--accent-glow);
+  border-color: var(--accent-border);
+  color: var(--accent);
+  box-shadow: 0 0 10px rgba(74, 222, 128, 0.18);
+}
+
+.task-done-btn:disabled {
+  opacity: 0.5;
+  cursor: wait;
+}
+
+/* Habit name */
+.task-name {
+  flex: 1;
+  font-size: 14px;
+  font-weight: 500;
+  color: var(--text);
+  min-width: 0;
+  transition: color var(--dur-fast), text-decoration-color var(--dur-fast);
+}
+
+.task-name.done {
+  color: var(--text-3);
+  text-decoration: line-through;
+  text-decoration-color: var(--text-3);
+}
+
+.task-description {
+  font-size: 11.5px;
+  color: var(--text-3);
+  margin-left: 8px;
+  font-weight: 400;
+}
+
+/* Streak pill */
+.task-streak {
+  font-size: 11.5px;
+  font-weight: 600;
+  color: var(--accent);
+  background: var(--accent-glow);
+  border: 1px solid var(--accent-border);
+  border-radius: 99px;
+  padding: 2px 9px;
+  white-space: nowrap;
+  flex-shrink: 0;
+}
+
+.task-streak.dim {
+  color: var(--text-3);
+  background: var(--surface-3);
+  border-color: var(--border-subtle);
+}
+
+/* Edit button — same pattern as delete, accent color on hover */
+.task-edit-btn {
+  width: 26px;
+  height: 26px;
+  border-radius: 7px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--text-3);
+  opacity: 0;
+  transition: opacity var(--dur-fast), background var(--dur-fast),
+    color var(--dur-fast);
+  flex-shrink: 0;
+}
+
+.task-row:hover .task-edit-btn {
+  opacity: 1;
+}
+
+.task-edit-btn:hover {
+  background: var(--accent-glow);
+  color: var(--accent);
+}
+
+/* Delete button */
+.task-delete-btn {
+  width: 26px;
+  height: 26px;
+  border-radius: 7px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--text-3);
+  opacity: 0;
+  transition: opacity var(--dur-fast), background var(--dur-fast),
+    color var(--dur-fast);
+  flex-shrink: 0;
+}
+
+.task-row:hover .task-delete-btn {
+  opacity: 1;
+}
+
+.task-delete-btn:hover {
+  background: var(--danger-bg);
+  color: var(--danger);
+}
+
+/* Divider */
+.task-divider {
+  height: 1px;
+  background: var(--border-subtle);
+  margin: 4px 10px;
+}
+
+/* Empty state */
+.tasks-empty {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 40px 20px;
+  color: var(--text-3);
+  gap: 10px;
+  text-align: center;
+}
+
+.tasks-empty-icon {
+  font-size: 32px;
+}
+
+.tasks-empty-text {
+  font-size: 13.5px;
+  color: var(--text-3);
+  line-height: 1.7;
+}
+
+.tasks-empty-text .accent {
+  color: var(--accent);
+}
