+++ frontend/src/components/layout/LoadingSpinner.jsx
@@ -1,0 +1,10 @@
+import "./LoadingSpinner.css";
+
+export default function LoadingSpinner() {
+  return (
+    <div className="loading-overlay">
+      <div className="spinner" />
+      Loading your habits…
+    </div>
+  );
+}
