+++ frontend/src/components/main/MainAppUI.css
@@ -1,0 +1,33 @@
+/* ============================================================
+   MAIN APP UI LAYOUT
+   ============================================================ */
+
+.app-shell {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  overflow: hidden;
+}
+
+.app-body {
+  flex: 1;
+  overflow-y: auto;
+  overflow-x: hidden;
+  padding: 24px 28px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.top-grid {
+  display: grid;
+  grid-template-columns: 1fr 360px;
+  gap: 20px;
+  min-height: 0;
+}
+
+.panel-header-count {
+  font-size: 12px;
+  color: var(--text-3);
+  font-weight: 500;
+}
