+++ frontend/src/components/shared/Panel.css
@@ -1,0 +1,44 @@
+/* ============================================================
+   SHARED PANEL STYLES
+   ============================================================ */
+
+.panel-card {
+  background: var(--surface);
+  border: 1px solid var(--border-subtle);
+  border-radius: var(--radius-lg);
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.panel-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 18px 22px 14px;
+  border-bottom: 1px solid var(--border-subtle);
+  flex-shrink: 0;
+}
+
+.panel-title {
+  font-family: var(--font-display);
+  font-size: 15px;
+  font-weight: 700;
+  color: var(--text);
+  letter-spacing: -0.01em;
+}
+
+.panel-title-sub {
+  font-size: 12px;
+  color: var(--text-3);
+  font-weight: 400;
+  margin-left: 8px;
+  font-family: var(--font-sans);
+  letter-spacing: 0;
+}
+
+.tasks-body {
+  flex: 1;
+  overflow-y: auto;
+  padding: 10px 12px;
+}
